+++ src/Workspace.cpp
@@ -9,7 +9,6 @@
 #include "glob/glob.hpp"
 #include "Luau/BuiltinDefinitions.h"
 
-LUAU_FASTFLAG(LuauStacklessTypeClone3)
 
 void WorkspaceFolder::openTextDocument(const lsp::DocumentUri& uri, const lsp::DidOpenTextDocumentParams& params)
 {
@@ -166,12 +165,6 @@ bool WorkspaceFolder::isDefinitionFile(const std::filesystem::path& path, const 
 // NOTE: use `frontend.parse` if you do not care about typechecking
 Luau::CheckResult WorkspaceFolder::checkSimple(const Luau::ModuleName& moduleName, bool runLintChecks)
 {
-    // TODO: We do not need to store the type graphs. But it leads to a bad bug if we disable it so for now, we keep the type graphs
-    // https://github.com/Roblox/luau/issues/975
-    if (!FFlag::LuauStacklessTypeClone3)
-        return frontend.check(
-            moduleName, Luau::FrontendOptions{/* retainFullTypeGraphs: */ true, /* forAutocomplete: */ false, /* runLintChecks: */ runLintChecks});
-
     try
     {
         return frontend.check(moduleName, Luau::FrontendOptions{/* retainFullTypeGraphs: */ false, /* forAutocomplete: */ false, runLintChecks});