+++ src/operations/Rename.cpp
@@ -3,6 +3,24 @@
 
 #include "LSP/LuauExt.hpp"
 
+static bool isGlobalBinding(const Luau::Frontend& frontend, const WorkspaceFileResolver& fileResolver, const lsp::RenameParams& params)
+{
+    auto moduleName = fileResolver.getModuleName(params.textDocument.uri);
+    auto textDocument = fileResolver.getTextDocument(params.textDocument.uri);
+    if (!textDocument)
+        throw JsonRpcException(lsp::ErrorCode::RequestFailed, "No managed text document for " + params.textDocument.uri.toString());
+    auto position = textDocument->convertPosition(params.position);
+
+    auto sourceModule = frontend.getSourceModule(moduleName);
+    // TODO: fix "forAutocomplete"
+    auto module = frontend.moduleResolverForAutocomplete.getModule(moduleName);
+    auto binding = Luau::findBindingAtPosition(*module, *sourceModule, position);
+    if (binding)
+        return binding->location.begin == Luau::Position{0, 0} && binding->location.end == Luau::Position{0, 0};
+
+    return false;
+}
+
 lsp::RenameResult WorkspaceFolder::rename(const lsp::RenameParams& params)
 {
     // Verify the new name is valid (is an identifier)
@@ -10,12 +28,10 @@ lsp::RenameResult WorkspaceFolder::rename(const lsp::RenameParams& params)
         throw JsonRpcException(lsp::ErrorCode::RequestFailed, "The new name must be a valid identifier");
     if (!isalpha(params.newName.at(0)) && params.newName.at(0) != '_')
         throw JsonRpcException(lsp::ErrorCode::RequestFailed, "The new name must be a valid identifier starting with a character or underscore");
-    for (auto ch : params.newName)
-    {
-        if (!isalpha(ch) && !isdigit(ch) && ch != '_')
-            throw JsonRpcException(
-                lsp::ErrorCode::RequestFailed, "The new name must be a valid identifier composed of characters, digits, and underscores only");
-    }
+
+    if (!Luau::isIdentifier(params.newName))
+        throw JsonRpcException(
+            lsp::ErrorCode::RequestFailed, "The new name must be a valid identifier composed of characters, digits, and underscores only");
 
     // Use findAllReferences to determine locations
     lsp::ReferenceParams referenceParams{};
@@ -26,6 +42,9 @@ lsp::RenameResult WorkspaceFolder::rename(const lsp::RenameParams& params)
 
     if (!references)
         throw JsonRpcException(lsp::ErrorCode::RequestFailed, "Unable to find symbol to rename");
+
+    if (isGlobalBinding(frontend, fileResolver, params))
+        throw JsonRpcException(lsp::ErrorCode::RequestFailed, "Cannot rename a global variable");
 
     lsp::WorkspaceEdit result{};