+++ src/platform/roblox/RobloxLuauExt.cpp
@@ -1,31 +1,11 @@
 #include "Platform/RobloxPlatform.hpp"
 
 #include "Luau/BuiltinDefinitions.h"
+#include "Luau/ConstraintSolver.h"
 #include "Luau/TypeInfer.h"
 
-// Since in Roblox land, debug is extended to introduce more methods, but the api-docs
-// mark the package name as `@luau` instead of `@roblox`
-static void fixDebugDocumentationSymbol(Luau::TypeId ty, const std::string& libraryName)
-{
-    auto mutableTy = Luau::asMutable(ty);
-    auto newDocumentationSymbol = mutableTy->documentationSymbol.value();
-    replace(newDocumentationSymbol, "@roblox", "@luau");
-    mutableTy->documentationSymbol = newDocumentationSymbol;
-
-    if (auto* ttv = Luau::getMutable<Luau::TableType>(ty))
-    {
-        ttv->name = "typeof(" + libraryName + ")";
-        for (auto& [name, prop] : ttv->props)
-        {
-            newDocumentationSymbol = prop.documentationSymbol.value();
-            replace(newDocumentationSymbol, "@roblox", "@luau");
-            prop.documentationSymbol = newDocumentationSymbol;
-        }
-    }
-}
-
 // Magic function for `Instance:IsA("ClassName")` predicate
-static std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionInstanceIsA(Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope,
+std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionInstanceIsA(Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope,
     const Luau::AstExprCall& expr, const Luau::WithPredicate<Luau::TypePackId>& withPredicate)
 {
     if (expr.args.size != 1)
@@ -55,8 +35,50 @@ static std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionInstanc
     return Luau::WithPredicate<Luau::TypePackId>{booleanPack, {Luau::IsAPredicate{std::move(*lvalue), expr.location, type}}};
 }
 
+static bool dcrMagicFunctionInstanceIsA(Luau::MagicFunctionCallContext context)
+{
+    if (context.callSite->args.size != 1)
+        return false;
+
+    auto index = context.callSite->func->as<Luau::AstExprIndexName>();
+    auto str = context.callSite->args.data[0]->as<Luau::AstExprConstantString>();
+    if (!index || !str)
+        return false;
+
+    std::string className(str->value.data, str->value.size);
+    std::optional<Luau::TypeFun> tfun = context.constraint->scope->lookupType(className);
+    if (!tfun)
+        context.solver->reportError(
+            Luau::TypeError{context.callSite->args.data[0]->location, Luau::UnknownSymbol{className, Luau::UnknownSymbol::Type}});
+
+    return false;
+}
+
+void dcrMagicRefinementInstanceIsA(const Luau::MagicRefinementContext& context)
+{
+    if (context.callSite->args.size != 1 || context.discriminantTypes.empty())
+        return;
+
+    auto index = context.callSite->func->as<Luau::AstExprIndexName>();
+    auto str = context.callSite->args.data[0]->as<Luau::AstExprConstantString>();
+    if (!index || !str)
+        return;
+
+    std::optional<Luau::TypeId> discriminantTy = context.discriminantTypes[0];
+    if (!discriminantTy)
+        return;
+
+    std::string className(str->value.data, str->value.size);
+    std::optional<Luau::TypeFun> tfun = context.scope->lookupType(className);
+    if (!tfun)
+        return;
+
+    LUAU_ASSERT(Luau::get<Luau::BlockedType>(*discriminantTy));
+    asMutable(*discriminantTy)->ty.emplace<Luau::BoundType>(Luau::follow(tfun->type));
+}
+
 // Magic function for `instance:Clone()`, so that we return the exact subclass that `instance` is, rather than just a generic Instance
-static std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionInstanceClone(Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope,
+std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionInstanceClone(Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope,
     const Luau::AstExprCall& expr, const Luau::WithPredicate<Luau::TypePackId>& withPredicate)
 {
     auto index = expr.func->as<Luau::AstExprIndexName>();
@@ -68,9 +90,24 @@ static std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionInstanc
     return Luau::WithPredicate<Luau::TypePackId>{arena.addTypePack({instanceType.type})};
 }
 
+static bool dcrMagicFunctionInstanceClone(Luau::MagicFunctionCallContext context)
+{
+    auto index = context.callSite->func->as<Luau::AstExprIndexName>();
+    if (!index)
+        return false;
+
+    // The cloned type is the self type, i.e. the first argument
+    auto selfTy = Luau::first(context.arguments);
+    if (!selfTy)
+        return false;
+
+    asMutable(context.result)->ty.emplace<Luau::BoundTypePack>(context.solver->arena->addTypePack({*selfTy}));
+    return true;
+}
+
 // Magic function for `Instance:FindFirstChildWhichIsA("ClassName")` and friends
-static std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionFindFirstXWhichIsA(Luau::TypeChecker& typeChecker,
-    const Luau::ScopePtr& scope, const Luau::AstExprCall& expr, const Luau::WithPredicate<Luau::TypePackId>& withPredicate)
+std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionFindFirstXWhichIsA(Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope,
+    const Luau::AstExprCall& expr, const Luau::WithPredicate<Luau::TypePackId>& withPredicate)
 {
     if (expr.args.size < 1)
         return std::nullopt;
@@ -90,8 +127,28 @@ static std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionFindFir
     return Luau::WithPredicate<Luau::TypePackId>{arena.addTypePack({nillableClass})};
 }
 
+static bool dcrMagicFunctionFindFirstXWhichIsA(Luau::MagicFunctionCallContext context)
+{
+    if (context.callSite->args.size < 1)
+        return false;
+
+    auto str = context.callSite->args.data[0]->as<Luau::AstExprConstantString>();
+    if (!str)
+        return false;
+
+    std::optional<Luau::TypeFun> tfun = context.constraint->scope->lookupType(std::string(str->value.data, str->value.size));
+    if (!tfun || !tfun->typeParams.empty() || !tfun->typePackParams.empty())
+        return false;
+
+    auto type = Luau::follow(tfun->type);
+
+    Luau::TypeId nillableClass = Luau::makeOption(context.solver->builtinTypes, *context.solver->arena, type);
+    asMutable(context.result)->ty.emplace<Luau::BoundTypePack>(context.solver->arena->addTypePack({nillableClass}));
+    return true;
+}
+
 // Magic function for `EnumItem:IsA("EnumType")` predicate
-static std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionEnumItemIsA(Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope,
+std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionEnumItemIsA(Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope,
     const Luau::AstExprCall& expr, const Luau::WithPredicate<Luau::TypePackId>& withPredicate)
 {
     if (expr.args.size != 1)
@@ -121,6 +178,48 @@ static std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionEnumIte
     return Luau::WithPredicate<Luau::TypePackId>{booleanPack, {Luau::IsAPredicate{std::move(*lvalue), expr.location, type}}};
 }
 
+static bool dcrMagicFunctionEnumItemIsA(Luau::MagicFunctionCallContext context)
+{
+    if (context.callSite->args.size != 1)
+        return false;
+
+    auto index = context.callSite->func->as<Luau::AstExprIndexName>();
+    auto str = context.callSite->args.data[0]->as<Luau::AstExprConstantString>();
+    if (!index || !str)
+        return false;
+
+    std::string enumItem(str->value.data, str->value.size);
+    std::optional<Luau::TypeFun> tfun = context.constraint->scope->lookupImportedType("Enum", enumItem);
+    if (!tfun)
+        context.solver->reportError(
+            Luau::TypeError{context.callSite->args.data[0]->location, Luau::UnknownSymbol{enumItem, Luau::UnknownSymbol::Type}});
+
+    return false;
+}
+
+static void dcrMagicRefinementEnumItemIsA(const Luau::MagicRefinementContext& context)
+{
+    if (context.callSite->args.size != 1 || context.discriminantTypes.empty())
+        return;
+
+    auto index = context.callSite->func->as<Luau::AstExprIndexName>();
+    auto str = context.callSite->args.data[0]->as<Luau::AstExprConstantString>();
+    if (!index || !str)
+        return;
+
+    std::optional<Luau::TypeId> discriminantTy = context.discriminantTypes[0];
+    if (!discriminantTy)
+        return;
+
+    std::string enumItem(str->value.data, str->value.size);
+    std::optional<Luau::TypeFun> tfun = context.scope->lookupImportedType("Enum", enumItem);
+    if (!tfun)
+        return;
+
+    LUAU_ASSERT(Luau::get<Luau::BlockedType>(*discriminantTy));
+    asMutable(*discriminantTy)->ty.emplace<Luau::BoundType>(Luau::follow(tfun->type));
+}
+
 // Magic function for `instance:GetPropertyChangedSignal()`, so that we can perform type checking on the provided property
 static std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionGetPropertyChangedSignal(Luau::TypeChecker& typeChecker,
     const Luau::ScopePtr& scope, const Luau::AstExprCall& expr, const Luau::WithPredicate<Luau::TypePackId>& withPredicate)
@@ -147,6 +246,129 @@ static std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionGetProp
     }
 
     return std::nullopt;
+}
+
+static bool dcrMagicFunctionGetPropertyChangedSignal(Luau::MagicFunctionCallContext context)
+{
+    if (context.callSite->args.size != 1)
+        return false;
+
+    auto index = context.callSite->func->as<Luau::AstExprIndexName>();
+    auto str = context.callSite->args.data[0]->as<Luau::AstExprConstantString>();
+    if (!index || !str)
+        return false;
+
+    // The cloned type is the self type, i.e. the first argument
+    auto selfTy = Luau::first(context.arguments);
+    if (!selfTy)
+        return false;
+
+    auto ctv = Luau::get<Luau::ClassType>(Luau::follow(selfTy));
+    if (!ctv)
+        return false;
+
+    std::string property(str->value.data, str->value.size);
+    if (!Luau::lookupClassProp(ctv, property))
+    {
+        context.solver->reportError(Luau::TypeError{context.callSite->args.data[0]->location, Luau::UnknownProperty{*selfTy, property}});
+    }
+
+    return false;
+}
+
+// Since in Roblox land, debug is extended to introduce more methods, but the api-docs
+// mark the package name as `@luau` instead of `@roblox`
+static void fixDebugDocumentationSymbol(Luau::TypeId ty, const std::string& libraryName)
+{
+    auto mutableTy = Luau::asMutable(ty);
+    auto newDocumentationSymbol = mutableTy->documentationSymbol.value();
+    replace(newDocumentationSymbol, "@roblox", "@luau");
+    mutableTy->documentationSymbol = newDocumentationSymbol;
+
+    if (auto* ttv = Luau::getMutable<Luau::TableType>(ty))
+    {
+        ttv->name = "typeof(" + libraryName + ")";
+        for (auto& [name, prop] : ttv->props)
+        {
+            newDocumentationSymbol = prop.documentationSymbol.value();
+            replace(newDocumentationSymbol, "@roblox", "@luau");
+            prop.documentationSymbol = newDocumentationSymbol;
+        }
+    }
+}
+
+static Luau::MagicFunction createMagicFunctionTypeLookup(const std::vector<std::string>& lookupList, const std::string& errorMessagePrefix)
+{
+    return [lookupList, errorMessagePrefix](Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope, const Luau::AstExprCall& expr,
+               const Luau::WithPredicate<Luau::TypePackId>& withPredicate) -> std::optional<Luau::WithPredicate<Luau::TypePackId>>
+    {
+        if (expr.args.size < 1)
+            return std::nullopt;
+
+        if (auto str = expr.args.data[0]->as<Luau::AstExprConstantString>())
+        {
+            auto className = std::string(str->value.data, str->value.size);
+            if (contains(lookupList, className))
+            {
+                std::optional<Luau::TypeFun> tfun = typeChecker.globalScope->lookupType(className);
+                if (!tfun || !tfun->typeParams.empty() || !tfun->typePackParams.empty())
+                {
+                    typeChecker.reportError(Luau::TypeError{expr.args.data[0]->location, Luau::UnknownSymbol{className, Luau::UnknownSymbol::Type}});
+                    return std::nullopt;
+                }
+
+                auto type = Luau::follow(tfun->type);
+
+                Luau::TypeArena& arena = typeChecker.currentModule->internalTypes;
+                Luau::TypePackId classTypePack = arena.addTypePack({type});
+                return Luau::WithPredicate<Luau::TypePackId>{classTypePack};
+            }
+            else
+            {
+                typeChecker.reportError(
+                    Luau::TypeError{expr.args.data[0]->location, Luau::GenericError{errorMessagePrefix + " '" + className + "'"}});
+            }
+        }
+
+        return std::nullopt;
+    };
+}
+
+static auto createDcrMagicFunctionTypeLookup(const std::vector<std::string>& lookupList, const std::string& errorMessagePrefix)
+{
+    return [lookupList, errorMessagePrefix](Luau::MagicFunctionCallContext context) -> bool
+    {
+        if (context.callSite->args.size < 1)
+            return false;
+
+        if (auto str = context.callSite->args.data[0]->as<Luau::AstExprConstantString>())
+        {
+            auto className = std::string(str->value.data, str->value.size);
+            if (contains(lookupList, className))
+            {
+                // TODO: only check the global scope?
+                std::optional<Luau::TypeFun> tfun = context.constraint->scope->lookupType(className);
+                if (!tfun || !tfun->typeParams.empty() || !tfun->typePackParams.empty())
+                {
+                    context.solver->reportError(
+                        Luau::TypeError{context.callSite->args.data[0]->location, Luau::UnknownSymbol{className, Luau::UnknownSymbol::Type}});
+                    return false;
+                }
+
+                auto type = Luau::follow(tfun->type);
+                Luau::TypePackId classTypePack = context.solver->arena->addTypePack({type});
+                asMutable(context.result)->ty.emplace<Luau::BoundTypePack>(classTypePack);
+                return true;
+            }
+            else
+            {
+                context.solver->reportError(
+                    Luau::TypeError{context.callSite->args.data[0]->location, Luau::GenericError{errorMessagePrefix + " '" + className + "'"}});
+            }
+        }
+
+        return false;
+    };
 }
 
 void RobloxPlatform::mutateRegisteredDefinitions(Luau::GlobalTypes& globals, std::optional<nlohmann::json> metadata)
@@ -181,6 +403,15 @@ void RobloxPlatform::mutateRegisteredDefinitions(Luau::GlobalTypes& globals, std
             Luau::attachMagicFunction(ctv->props["FindFirstAncestorOfClass"].type(), magicFunctionFindFirstXWhichIsA);
             Luau::attachMagicFunction(ctv->props["Clone"].type(), magicFunctionInstanceClone);
             Luau::attachMagicFunction(ctv->props["GetPropertyChangedSignal"].type(), magicFunctionGetPropertyChangedSignal);
+
+            Luau::attachDcrMagicRefinement(ctv->props["IsA"].type(), dcrMagicRefinementInstanceIsA);
+            Luau::attachDcrMagicFunction(ctv->props["IsA"].type(), dcrMagicFunctionInstanceIsA);
+            Luau::attachDcrMagicFunction(ctv->props["FindFirstChildWhichIsA"].type(), dcrMagicFunctionFindFirstXWhichIsA);
+            Luau::attachDcrMagicFunction(ctv->props["FindFirstChildOfClass"].type(), dcrMagicFunctionFindFirstXWhichIsA);
+            Luau::attachDcrMagicFunction(ctv->props["FindFirstAncestorWhichIsA"].type(), dcrMagicFunctionFindFirstXWhichIsA);
+            Luau::attachDcrMagicFunction(ctv->props["FindFirstAncestorOfClass"].type(), dcrMagicFunctionFindFirstXWhichIsA);
+            Luau::attachDcrMagicFunction(ctv->props["Clone"].type(), dcrMagicFunctionInstanceClone);
+            Luau::attachDcrMagicFunction(ctv->props["GetPropertyChangedSignal"].type(), dcrMagicFunctionGetPropertyChangedSignal);
 
             // Autocomplete ClassNames for :IsA("") and counterparts
             Luau::attachTag(ctv->props["IsA"].type(), "ClassNames");
@@ -222,7 +453,9 @@ void RobloxPlatform::mutateRegisteredDefinitions(Luau::GlobalTypes& globals, std
 
                     Luau::attachTag(newFunction->second.type(), "CreatableInstances");
                     Luau::attachMagicFunction(
-                        newFunction->second.type(), types::createMagicFunctionTypeLookup(robloxMetadata->CREATABLE_INSTANCES, "Invalid class name"));
+                        newFunction->second.type(), createMagicFunctionTypeLookup(robloxMetadata->CREATABLE_INSTANCES, "Invalid class name"));
+                    Luau::attachDcrMagicFunction(
+                        newFunction->second.type(), createDcrMagicFunctionTypeLookup(robloxMetadata->CREATABLE_INSTANCES, "Invalid class name"));
                 }
 
     // Attach onto `game:GetService()`
@@ -233,7 +466,9 @@ void RobloxPlatform::mutateRegisteredDefinitions(Luau::GlobalTypes& globals, std
             {
                 Luau::attachTag(ctv->props["GetService"].type(), "Services");
                 Luau::attachMagicFunction(
-                    ctv->props["GetService"].type(), types::createMagicFunctionTypeLookup(robloxMetadata->SERVICES, "Invalid service name"));
+                    ctv->props["GetService"].type(), createMagicFunctionTypeLookup(robloxMetadata->SERVICES, "Invalid service name"));
+                Luau::attachDcrMagicFunction(
+                    ctv->props["GetService"].type(), createDcrMagicFunctionTypeLookup(robloxMetadata->SERVICES, "Invalid service name"));
             }
 
     // Move Enums over as imported type bindings
@@ -249,6 +484,8 @@ void RobloxPlatform::mutateRegisteredDefinitions(Luau::GlobalTypes& globals, std
                 if (ctv->name == "EnumItem")
                 {
                     Luau::attachMagicFunction(ctv->props["IsA"].type(), magicFunctionEnumItemIsA);
+                    Luau::attachDcrMagicFunction(ctv->props["IsA"].type(), dcrMagicFunctionEnumItemIsA);
+                    Luau::attachDcrMagicRefinement(ctv->props["IsA"].type(), dcrMagicRefinementEnumItemIsA);
                     Luau::attachTag(ctv->props["IsA"].type(), "Enums");
                 }
                 else if (ctv->name != "Enum" && ctv->name != "Enums")