+++ tests/Hover.test.cpp
@@ -37,4 +37,123 @@ TEST_CASE_FIXTURE(Fixture, "show_string_utf8_characters_on_hover")
     CHECK_EQ(result->contents.value, codeBlock("luau", "string (22 bytes, 19 characters)"));
 }
 
+TEST_CASE_FIXTURE(Fixture, "basic_type_alias_declaration")
+{
+    auto source = R"(
+        type Identity = string
+    )";
+
+    auto uri = newDocument("foo.luau", source);
+
+    lsp::HoverParams params;
+    params.textDocument = lsp::TextDocumentIdentifier{uri};
+    params.position = lsp::Position{1, 18};
+
+    auto result = workspace.hover(params);
+    REQUIRE(result);
+    CHECK_EQ(result->contents.value, codeBlock("luau", "type Identity = string"));
+}
+
+TEST_CASE_FIXTURE(Fixture, "type_alias_declaration_with_single_generic")
+{
+    auto source = R"(
+        type Identity<T> = T
+    )";
+
+    auto uri = newDocument("foo.luau", source);
+
+    lsp::HoverParams params;
+    params.textDocument = lsp::TextDocumentIdentifier{uri};
+    params.position = lsp::Position{1, 18};
+
+    auto result = workspace.hover(params);
+    REQUIRE(result);
+    CHECK_EQ(result->contents.value, codeBlock("luau", "type Identity<T> = T"));
+}
+
+TEST_CASE_FIXTURE(Fixture, "type_alias_declaration_with_generic_default_value")
+{
+    auto source = R"(
+        type Identity<T = string> = T
+    )";
+
+    auto uri = newDocument("foo.luau", source);
+
+    lsp::HoverParams params;
+    params.textDocument = lsp::TextDocumentIdentifier{uri};
+    params.position = lsp::Position{1, 18};
+
+    auto result = workspace.hover(params);
+    REQUIRE(result);
+    CHECK_EQ(result->contents.value, codeBlock("luau", "type Identity<T = string> = T"));
+}
+
+TEST_CASE_FIXTURE(Fixture, "type_alias_declaration_with_multiple_generics")
+{
+    auto source = R"(
+        type Identity<T, U = string> = T
+    )";
+
+    auto uri = newDocument("foo.luau", source);
+
+    lsp::HoverParams params;
+    params.textDocument = lsp::TextDocumentIdentifier{uri};
+    params.position = lsp::Position{1, 18};
+
+    auto result = workspace.hover(params);
+    REQUIRE(result);
+    CHECK_EQ(result->contents.value, codeBlock("luau", "type Identity<T, U = string> = T"));
+}
+
+TEST_CASE_FIXTURE(Fixture, "type_alias_declaration_generic_type_pack")
+{
+    auto source = R"(
+        type Identity<T...> = (any) -> T...
+    )";
+
+    auto uri = newDocument("foo.luau", source);
+
+    lsp::HoverParams params;
+    params.textDocument = lsp::TextDocumentIdentifier{uri};
+    params.position = lsp::Position{1, 18};
+
+    auto result = workspace.hover(params);
+    REQUIRE(result);
+    CHECK_EQ(result->contents.value, codeBlock("luau", "type Identity<T...> = (any) -> (T...)"));
+}
+
+TEST_CASE_FIXTURE(Fixture, "type_alias_declaration_generic_type_pack_with_default")
+{
+    auto source = R"(
+        type Identity<T... = ...string> = (any) -> T...
+    )";
+
+    auto uri = newDocument("foo.luau", source);
+
+    lsp::HoverParams params;
+    params.textDocument = lsp::TextDocumentIdentifier{uri};
+    params.position = lsp::Position{1, 18};
+
+    auto result = workspace.hover(params);
+    REQUIRE(result);
+    CHECK_EQ(result->contents.value, codeBlock("luau", "type Identity<T... = ...string> = (any) -> (T...)"));
+}
+
+TEST_CASE_FIXTURE(Fixture, "complex_type_alias_declaration_with_generics")
+{
+    auto source = R"(
+        type Identity<T, U = number, V... = ...string> = (T, U) -> V...
+    )";
+
+    auto uri = newDocument("foo.luau", source);
+
+    lsp::HoverParams params;
+    params.textDocument = lsp::TextDocumentIdentifier{uri};
+    params.position = lsp::Position{1, 18};
+
+    auto result = workspace.hover(params);
+    REQUIRE(result);
+    CHECK_EQ(result->contents.value, codeBlock("luau", "type Identity<T, U = number, V... = ...string> = (T, U) -> (V...)"));
+}
+
 TEST_SUITE_END();